Fixing the redirect logic for preview apps for signup
